move env vars out of `pkg`
