+++ pkg/cosign/verify.go
@@ -150,13 +150,13 @@ type CheckOpts struct {
 
 // Verify does all the main cosign checks in a loop, returning validated payloads.
 // If there were no payloads, we return an error.
-func Verify(ctx context.Context, ref name.Reference, co *CheckOpts) ([]SignedPayload, error) {
+func Verify(ctx context.Context, ref name.Reference, co *CheckOpts, rekorServerURL string) ([]SignedPayload, error) {
 	// Enforce this up front.
 	if co.Roots == nil && co.PubKey == nil {
 		return nil, errors.New("one of public key or cert roots is required")
 	}
 	// TODO: Figure out if we'll need a client before creating one.
-	rekorClient, err := app.GetRekorClient(TlogServer())
+	rekorClient, err := app.GetRekorClient(rekorServerURL)
 	if err != nil {
 		return nil, err
 	}
